healer fix

+++ src/main/java/noppes/npcs/controllers/DialogController.java
@@ -2,10 +2,7 @@ package noppes.npcs.controllers;
 
 import java.io.File;
 import java.nio.file.Files;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Objects;
-import java.util.TreeMap;
+import java.util.*;
 
 import net.minecraft.nbt.CompressedStreamTools;
 import net.minecraft.nbt.NBTTagCompound;
@@ -17,7 +14,6 @@ import noppes.npcs.NoppesStringUtils;
 import noppes.npcs.Server;
 import noppes.npcs.api.constants.OptionType;
 import noppes.npcs.api.handler.IDialogHandler;
-import noppes.npcs.api.handler.data.IDialog;
 import noppes.npcs.api.handler.data.IDialogCategory;
 import noppes.npcs.constants.EnumPacketClient;
 import noppes.npcs.constants.EnumSync;
@@ -45,11 +41,11 @@ public class DialogController implements IDialogHandler {
 
 	@Override
 	public IDialogCategory[] categories() {
-		return this.categories.values().toArray(new IDialogCategory[0]);
+		return categories.values().toArray(new IDialogCategory[0]);
 	}
 
 	public boolean containsCategoryName(DialogCategory category) {
-		for (DialogCategory cat : this.categories.values()) {
+		for (DialogCategory cat : categories.values()) {
 			if (category.id != cat.id && cat.title.equalsIgnoreCase(category.title)) {
 				return true;
 			}
@@ -67,22 +63,20 @@ public class DialogController implements IDialogHandler {
 	}
 
 	@Override
-	public IDialog get(int id) {
-		return this.dialogs.get(id);
-	}
+	public Dialog get(int id) { return dialogs.get(id); }
 
 	private File getDir() {
 		return new File(CustomNpcs.getWorldSaveDirectory(), "dialogs");
 	}
 
 	public boolean hasDialog(int dialogId) {
-		return this.dialogs.containsKey(dialogId);
+		return dialogs.containsKey(dialogId);
 	}
 
 	public void load() {
 		CustomNpcs.debugData.start(null);
 		LogWriter.info("Loading Dialogs");
-		this.loadCategories();
+		loadCategories();
 		try {
 			File file = new File(CustomNpcs.getWorldSaveDirectory(), "dialog_gui_settings.dat");
 			if (file.exists()) { guiSettings.load(CompressedStreamTools.readCompressed(Files.newInputStream(file.toPath()))); }
@@ -94,49 +88,46 @@ public class DialogController implements IDialogHandler {
 
 	@SuppressWarnings("all")
 	private void loadCategories() {
-		this.categories.clear();
-		this.dialogs.clear();
-		this.lastUsedCatID = 0;
-		this.lastUsedDialogID = 0;
+		categories.clear();
+		dialogs.clear();
+		lastUsedCatID = 0;
+		lastUsedDialogID = 0;
 		try {
 			File file = new File(CustomNpcs.getWorldSaveDirectory(), "dialog.dat");
 			if (file.exists()) {
-				this.loadCategoriesOld(file);
+				loadCategoriesOld(file);
 				file.delete();
 				file = new File(CustomNpcs.getWorldSaveDirectory(), "dialog.dat_old");
-				if (file.exists()) {
-					file.delete();
-				}
+				if (file.exists()) { file.delete(); }
 				return;
 			}
 		} catch (Exception e) { LogWriter.except(e); }
-		File dir = this.getDir();
+		File dir = getDir();
 		if (!dir.exists()) {
 			dir.mkdir();
-			this.loadDefaultDialogs();
+			loadDefaultDialogs();
 		}
 		else {
-			for (File file2 : Objects.requireNonNull(dir.listFiles())) {
-				if (file2.isDirectory()) {
-					DialogCategory category = this.loadCategoryDir(file2);
-					Iterator<Map.Entry<Integer, Dialog>> ite = category.dialogs.entrySet().iterator();
-					while (ite.hasNext()) {
-						Map.Entry<Integer, Dialog> entry = ite.next();
-						int id = entry.getKey();
-						if (id > this.lastUsedDialogID) {
-							this.lastUsedDialogID = id;
+			File[] files = dir.listFiles();
+			if (files != null) {
+				for (File dialogFile : files) {
+					if (dialogFile.isDirectory()) {
+						DialogCategory category = loadCategoryDir(dialogFile);
+						for (Map.Entry<Integer, Dialog> entry : new ArrayList<>(category.dialogs.entrySet())) {
+							Integer id = entry.getKey();
+							if (id > lastUsedDialogID) { lastUsedDialogID = id; }
+							Dialog dialog = entry.getValue();
+							if (dialogs.containsKey(id)) {
+								LogWriter.error("Duplicate dialog ID:" + dialog.id + " from category " + category.title);
+								category.dialogs.remove(id);
+							} else {
+								dialogs.put(id, dialog);
+							}
 						}
-						Dialog dialog = entry.getValue();
-						if (this.dialogs.containsKey(id)) {
-							LogWriter.error("Duplicate id " + dialog.id + " from category " + category.title);
-							ite.remove();
-						} else {
-							this.dialogs.put(id, dialog);
-						}
+						++lastUsedCatID;
+						category.id = lastUsedCatID;
+						categories.put(category.id, category);
 					}
-					++this.lastUsedCatID;
-					category.id = this.lastUsedCatID;
-					this.categories.put(category.id, category);
 				}
 			}
 		}
@@ -147,18 +138,15 @@ public class DialogController implements IDialogHandler {
 		NBTTagList list = compound.getTagList("Data", 10);
         for (int i = 0; i < list.tagCount(); ++i) {
 			DialogCategory category = new DialogCategory();
-			category.readNBT(list.getCompoundTagAt(i));
-			this.saveCategory(category);
+			category.load(list.getCompoundTagAt(i));
+			saveCategory(category);
 			Iterator<Map.Entry<Integer, Dialog>> ita = category.dialogs.entrySet().iterator();
 			while (ita.hasNext()) {
 				Map.Entry<Integer, Dialog> entry = ita.next();
 				Dialog dialog = entry.getValue();
 				dialog.id = entry.getKey();
-				if (this.dialogs.containsKey(dialog.id)) {
-					ita.remove();
-				} else {
-					this.saveDialog(category, dialog);
-				}
+				if (dialogs.containsKey(dialog.id)) { ita.remove(); }
+				else { saveDialog(category, dialog); }
 			}
 		}
 	}
@@ -166,13 +154,14 @@ public class DialogController implements IDialogHandler {
 	private DialogCategory loadCategoryDir(File dir) {
 		DialogCategory category = new DialogCategory();
 		category.title = dir.getName();
-		for (File file : Objects.requireNonNull(dir.listFiles())) {
-			if (file.isFile()) {
-				if (file.getName().endsWith(".json")) {
+		File[] files = dir.listFiles();
+		if (files != null) {
+			for (File file : files) {
+				if (file.isFile() && file.getName().endsWith(".json")) {
 					try {
 						Dialog dialog = new Dialog(category);
 						dialog.id = Integer.parseInt(file.getName().substring(0, file.getName().length() - 5));
-						dialog.readNBTPartial(NBTJsonUtil.LoadFile(file));
+						dialog.loadPartial(NBTJsonUtil.LoadFile(file));
 						category.dialogs.put(dialog.id, dialog);
 					} catch (Exception e) {
 						LogWriter.error("Error loading: " + file.getAbsolutePath(), e);
@@ -185,77 +174,76 @@ public class DialogController implements IDialogHandler {
 
 	private void loadDefaultDialogs() {
 		DialogCategory cat = new DialogCategory();
-		cat.id = this.lastUsedCatID++;
+		cat.id = lastUsedCatID++;
 		cat.title = "Villager";
 		Dialog dia1 = new Dialog(cat);
 		dia1.id = 1;
 		dia1.title = "Start";
 		dia1.text = "dialog.base.1.text";
+
 		Dialog dia2 = new Dialog(cat);
 		dia2.id = 2;
 		dia2.title = "Ask about village";
-		dia2.text = Util.instance
-				.deleteColor(new TextComponentTranslation("dialog.base.2.text").getFormattedText());
+		dia2.text = Util.instance.deleteColor(new TextComponentTranslation("dialog.base.2.text").getFormattedText());
+
 		Dialog dia3 = new Dialog(cat);
 		dia3.id = 3;
 		dia3.title = "Who are you";
-		dia3.text = Util.instance
-				.deleteColor(new TextComponentTranslation("dialog.base.3.text").getFormattedText());
+		dia3.text = Util.instance.deleteColor(new TextComponentTranslation("dialog.base.3.text").getFormattedText());
+
 		cat.dialogs.put(dia1.id, dia1);
 		cat.dialogs.put(dia2.id, dia2);
 		cat.dialogs.put(dia3.id, dia3);
+
 		DialogOption option = new DialogOption();
 		option.title = "dialog.base.1.option.0";
 		option.addDialog(2);
 		option.optionType = OptionType.DIALOG_OPTION;
+
 		DialogOption option2 = new DialogOption();
 		option2.title = "dialog.base.1.option.1";
 		option2.addDialog(3);
 		option2.optionType = OptionType.DIALOG_OPTION;
+
 		DialogOption option3 = new DialogOption();
 		option3.title = "dialog.base.1.option.2";
 		option3.optionType = OptionType.QUIT_OPTION;
 		dia1.options.put(0, option2);
 		dia1.options.put(1, option);
 		dia1.options.put(2, option3);
+
 		DialogOption option4 = new DialogOption();
 		option4.title = Util.instance.deleteColor(new TextComponentTranslation("dialog.base.2.option.0").getFormattedText());
 		option4.addDialog(1);
 		dia2.options.put(1, option4);
 		dia3.options.put(1, option4);
-		this.lastUsedDialogID = 3;
-		this.lastUsedCatID = 1;
-		this.saveCategory(cat);
-		this.saveDialog(cat, dia1);
-		this.saveDialog(cat, dia2);
-		this.saveDialog(cat, dia3);
+
+		lastUsedDialogID = 3;
+		lastUsedCatID = 1;
+		saveCategory(cat);
+		saveDialog(cat, dia1);
+		saveDialog(cat, dia2);
+		saveDialog(cat, dia3);
 	}
 
 	public void removeCategory(int category) {
-		DialogCategory cat = this.categories.get(category);
-		if (cat == null) {
-			return;
-		}
-		File dir = new File(this.getDir(), cat.title);
-		if (!Util.instance.removeFile(dir)) {
-			LogWriter.error("Error delete " + dir + "; no access or file not uploaded!");
-		}
-		for (int dia : cat.dialogs.keySet()) {
-			this.dialogs.remove(dia);
-		}
-		this.categories.remove(category);
+		DialogCategory cat = categories.get(category);
+		if (cat == null) { return; }
+		File dir = new File(getDir(), cat.title);
+		if (!Util.instance.removeFile(dir)) { LogWriter.error("Error delete " + dir + "; no access or file not uploaded!"); }
+		for (int dia : cat.dialogs.keySet()) { dialogs.remove(dia); }
+		categories.remove(category);
 		Server.sendToAll(CustomNpcs.Server, EnumPacketClient.SYNC_REMOVE, EnumSync.DialogCategoriesData, category);
 	}
 
 	public void removeDialog(Dialog dialog) {
 		DialogCategory category = dialog.category;
-		File file = new File(new File(this.getDir(), category.title), dialog.id + ".json");
-		if (!file.delete()) {
-			return;
+		File file = new File(new File(getDir(), category.title), dialog.id + ".json");
+		if (file.delete()) {
+			category.dialogs.remove(dialog.id);
+			dialogs.remove(dialog.id);
+			Server.sendToAll(CustomNpcs.Server, EnumPacketClient.SYNC_REMOVE, EnumSync.DialogData, dialog.id);
 		}
-		category.dialogs.remove(dialog.id);
-		this.dialogs.remove(dialog.id);
-		Server.sendToAll(CustomNpcs.Server, EnumPacketClient.SYNC_REMOVE, EnumSync.DialogData, dialog.id);
 	}
 
 	@SuppressWarnings("all")
@@ -264,41 +252,41 @@ public class DialogController implements IDialogHandler {
 		category.title = NoppesStringUtils.cleanFileName(category.title);
 		if (category.title.isEmpty()) {
 			category.title = "default";
-			while (this.containsCategoryName(category)) { category.title += "_"; }
+			while (containsCategoryName(category)) { category.title += "_"; }
 		}
-		if (this.categories.containsKey(category.id)) {
-			DialogCategory currentCategory = this.categories.get(category.id);
+		if (categories.containsKey(category.id)) {
+			DialogCategory currentCategory = categories.get(category.id);
 			if (!currentCategory.title.equals(category.title)) {
-				while (this.containsCategoryName(category)) { category.title += "_"; }
-				File newDir = new File(this.getDir(), category.title);
-				File oldDir = new File(this.getDir(), currentCategory.title);
-				if (newDir.exists() || !oldDir.renameTo(newDir)) {
+				while (containsCategoryName(category)) { category.title += "_"; }
+				File newDir = new File(getDir(), category.title);
+				File oldDir = new File(getDir(), currentCategory.title);
+				if (newDir.exists()) {
+					CustomNpcs.debugData.end(null);
+					if (oldDir.exists()) { Util.instance.removeFile(oldDir); }
+					return;
+				}
+				else if (!oldDir.renameTo(newDir)) {
 					CustomNpcs.debugData.end(null);
 					return;
 				}
 			}
 			category.dialogs.clear();
 			category.dialogs.putAll(currentCategory.dialogs);
-		} else {
+		}
+		else {
 			if (category.id < 0) {
-				++this.lastUsedCatID;
-				category.id = this.lastUsedCatID;
+				++lastUsedCatID;
+				category.id = lastUsedCatID;
 			}
-			while (this.containsCategoryName(category)) {
-				category.title += "_";
-			}
-			File dir = new File(this.getDir(), category.title);
-			if (!dir.exists()) {
-				dir.mkdirs();
-			}
+			while (containsCategoryName(category)) { category.title += "_"; }
+			File dir = new File(getDir(), category.title);
+			if (!dir.exists()) { dir.mkdirs(); }
 		}
-		this.categories.put(category.id, category);
+		categories.put(category.id, category);
 		for (Dialog dialog : dialogs.values()) {
-			if (dialog.category.id == category.id) {
-				dialog.category = category;
-			}
+			if (dialog.category.id == category.id) { dialog.category = category; }
 		}
-		Server.sendToAll(CustomNpcs.Server, EnumPacketClient.SYNC_UPDATE, EnumSync.DialogCategoriesData, category.writeNBT(new NBTTagCompound()));
+		Server.sendToAll(CustomNpcs.Server, EnumPacketClient.SYNC_UPDATE, EnumSync.DialogCategoriesData, category.save(new NBTTagCompound()));
 		CustomNpcs.debugData.end(null);
 	}
 
@@ -306,32 +294,24 @@ public class DialogController implements IDialogHandler {
 	public void saveDialog(DialogCategory category, Dialog dialog) {
 		if (category == null) { return; }
 		CustomNpcs.debugData.start(null);
-		StringBuilder title = new StringBuilder(dialog.title);
-		while (this.containsDialogName(dialog.category, dialog)) {
-			title.append("_");
-		}
-		dialog.title = title.toString();
+		while(containsDialogName(dialog.category, dialog)) { dialog.title = dialog.title + "_"; }
 		if (dialog.id < 0) {
-			++this.lastUsedDialogID;
-			dialog.id = this.lastUsedDialogID;
+			++lastUsedDialogID;
+			dialog.id = lastUsedDialogID;
 		}
-		this.dialogs.put(dialog.id, dialog);
+		dialogs.put(dialog.id, dialog);
 		category.dialogs.put(dialog.id, dialog);
-		File dir = new File(this.getDir(), category.title);
-		if (!dir.exists()) {
-			dir.mkdirs();
-		}
+
+		File dir = new File(getDir(), category.title);
+		if (!dir.exists()) { dir.mkdirs(); }
 		File file = new File(dir, dialog.id + ".json_new");
 		File file2 = new File(dir, dialog.id + ".json");
 		try {
-			NBTTagCompound compound = dialog.writeToNBT(new NBTTagCompound());
+			NBTTagCompound compound = dialog.save(new NBTTagCompound());
 			Util.instance.saveFile(file, compound);
-			if (file2.exists()) {
-				file2.delete();
-			}
+			if (file2.exists()) { file2.delete(); }
 			file.renameTo(file2);
-			Server.sendToAll(CustomNpcs.Server, EnumPacketClient.SYNC_UPDATE, EnumSync.DialogData, compound,
-					category.id);
+			Server.sendToAll(CustomNpcs.Server, EnumPacketClient.SYNC_UPDATE, EnumSync.DialogData, compound, category.id);
 		}
 		catch (Exception e) { LogWriter.except(e);}
 		CustomNpcs.debugData.end(null);
